requirements.txt
django
django-mptt
djangorestframework
psycopg2
pymongo
requests
xmltodict